chatbot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
chromadb==0.5.5
requests==2.32.3
pypdf==4.3.1
python-docx==1.1.2
python-multipart==0.0.9
python-dotenv==1.0.1
